Refactor: Update AppTheme import path
