+++ app/src/main/java/com/washingtondcsquad/tudee/presentation/screen/onBoarding/OnBoardingButton.kt
@@ -16,8 +16,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.washingtondcsquad.tudee.R
-import com.washingtondcsquad.tudee.presentation.design.theme.AppTheme
-import com.washingtondcsquad.tudee.presentation.design.theme.TudeeTheme
+import com.washingtondcsquad.tudee.presentation.design.AppTheme
 import com.washingtondcsquad.tudee.presentation.utils.modifierExensions.dropShadow
 
 
@@ -56,7 +55,7 @@ fun OnBoardingButton(modifier: Modifier = Modifier, onClick: () -> Unit) {
 @Preview
 @Composable
 private fun OnBoardingButtonPrev() {
-    TudeeTheme(true) {
+    AppTheme(true) {
         OnBoardingButton(onClick = {})
     }
 }
